requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
pyyaml>=6.0
jinja2>=3.1
python-dateutil>=2.9
gql>=3.5
